requirements.txt
Django==4.2.28
djangorestframework==3.16.1
psycopg2==2.9.11
python-decouple==3.8